feat:#55 모든 채팅을 한번에 불러오도록 수정
